[python] Move np and rotate to utils
